[HOTFIX] UsedCar Structure Hot Fix + OneToOne Fetch Add + DeleteMapping -> ResponseEntity
