+++ src/main/kotlin/sundaystudy/kr/usedcar/module/usedcar/dto/request/InsuranceRequest.kt
@@ -1,5 +1,5 @@
-package sundaystudy.kr.usedcar.module.usedcar.insurance.dto.request
-import sundaystudy.kr.usedcar.module.usedcar.insurance.dto.detail.*
+package sundaystudy.kr.usedcar.module.usedcar.dto.request
+import sundaystudy.kr.usedcar.module.usedcar.dto.detail.*
 import java.util.*
 
 data class InsuranceRequest(
