+++ src/main/kotlin/sundaystudy/kr/usedcar/module/usedcar/dto/request/UsedCarSaveRequest.kt
@@ -1,6 +1,4 @@
 package sundaystudy.kr.usedcar.module.usedcar.dto.request
-import sundaystudy.kr.usedcar.module.usedcar.insurance.dto.request.InsuranceRequest
-import java.util.UUID
 
 data class UsedCarSaveRequest(
     var licenseNumber: String,
